+++ packages/core/src/tracing/instrumentation/protocols/_http.js
@@ -1,5 +1,9 @@
 'use strict';
 
+/*
+ * Used in http server instrumentation, the headers are already lower cased in the data we receive. This method is less
+ * expensive than getExtraHeadersCaseInsensitive.
+ */
 exports.getExtraHeaders = function getExtraHeaders(message, extraHttpHeadersToCapture) {
   if (!extraHttpHeadersToCapture || extraHttpHeadersToCapture.length === 0) {
     return undefined;
@@ -21,11 +25,63 @@ exports.getExtraHeaders = function getExtraHeaders(message, extraHttpHeadersToCa
   return extraHeaders;
 };
 
-exports.mergeExtraHeaders = function mergeExtraHeaders(
+/*
+ * Used in http client instrumentation, the headers can appear in any lower case/upper case combination there. This is
+ * slightly more expensive than getExtraHeaders.
+ */
+exports.getExtraHeadersCaseInsensitive = function getExtraHeaders(options, extraHttpHeadersToCapture) {
+  if (
+    !extraHttpHeadersToCapture ||
+    extraHttpHeadersToCapture.length === 0 ||
+    !options ||
+    !options.headers ||
+    typeof options.headers !== 'object'
+  ) {
+    return undefined;
+  }
+
+  var keys = Object.keys(options.headers).map(function(key) {
+    return { orig: key, low: key.toLowerCase() };
+  });
+  var extraHeadersFound = false;
+  var extraHeaders = {};
+  for (var i = 0; i < extraHttpHeadersToCapture.length; i++) {
+    var keyToCapture = extraHttpHeadersToCapture[i];
+    for (var j = 0; j < keys.length; j++) {
+      if (keys[j].low === keyToCapture) {
+        extraHeaders[keys[j].low] = options.headers[keys[j].orig];
+        extraHeadersFound = true;
+      }
+    }
+  }
+  if (!extraHeadersFound) {
+    return undefined;
+  }
+  return extraHeaders;
+};
+
+// eslint-disable-next-line max-len
+exports.mergeExtraHeadersFromServerResponseOrClientResponse = function mergeExtraHeadersFromServerResponseOrClientResponse(
   headersAlreadyCapturedIfAny,
-  message,
+  serverResponse,
   extraHttpHeadersToCapture
 ) {
+  return mergeExtraHeaders(headersAlreadyCapturedIfAny, extraHttpHeadersToCapture, function(key) {
+    return serverResponse.getHeader(key);
+  });
+};
+
+exports.mergeExtraHeadersFromIncomingMessage = function mergeExtraHeadersForExit(
+  headersAlreadyCapturedIfAny,
+  incomingMessage,
+  extraHttpHeadersToCapture
+) {
+  return mergeExtraHeaders(headersAlreadyCapturedIfAny, extraHttpHeadersToCapture, function(key) {
+    return incomingMessage.headers[key];
+  });
+};
+
+function mergeExtraHeaders(headersAlreadyCapturedIfAny, extraHttpHeadersToCapture, getHeader) {
   if (!extraHttpHeadersToCapture || extraHttpHeadersToCapture.length === 0) {
     return headersAlreadyCapturedIfAny;
   }
@@ -34,7 +90,7 @@ exports.mergeExtraHeaders = function mergeExtraHeaders(
   var additionalHeaders = {};
   for (var i = 0; i < extraHttpHeadersToCapture.length; i++) {
     var key = extraHttpHeadersToCapture[i];
-    var value = message.getHeader(key);
+    var value = getHeader(key);
     if (value) {
       additionalHeaders[key] = value;
       additionalHeadersFound = true;
@@ -50,4 +106,4 @@ exports.mergeExtraHeaders = function mergeExtraHeaders(
     return additionalHeaders;
   }
   return headersAlreadyCapturedIfAny;
-};
+}